Address review with felixfontein: doc updates and module splitting. Added "created" state.
